+++ CHANGELOG.md
@@ -7,6 +7,9 @@ First release.
 > This section will be removed after the beta phase. <br>
 > Note that semantic versioning rules are not strictly followed during this phase.
 
+- v0.2.1: new option `tree.options.emptyChildListExpandable` controls if
+  parent nodes with `node.children === []` are still expandable 
+  (like macOS Finder). Defaults to false.
 - v0.2.1: `load()`, `source` and `lazyLoad` now support an extended sytax :
   ```js
   source: {